% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaFortR-package.R
\docType{package}
\name{AquaFortR-package}
\alias{AquaFortR}
\alias{AquaFortR-package}
\title{AquaFortR: Streamlining Atmospheric Science, Oceanography, Climate, and Water Research with Fortran-accelerated R}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://AHomoudi.github.io/AquaFortR/}
}

}
\author{
\strong{Maintainer}: Ahmed Homoudi \email{ahmed.homoudi@tu-dresden.de} (\href{https://orcid.org/0000-0001-6906-5986}{ORCID})

}
\keyword{internal}
