% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cape_f.R
\name{cape_f}
\alias{cape_f}
\title{CAPE using Fortran}
\usage{
cape_f(
  t_parcel,
  dwpt_parcel,
  mr_parcel,
  p_profile,
  t_profile,
  mr_profile,
  vtc = TRUE
)
}
\arguments{
\item{t_parcel}{Parcel temperature [K].}

\item{dwpt_parcel}{Parcel dew point temperature [K].}

\item{mr_parcel}{Parcel mixing ratio [kg/kg].}

\item{p_profile}{Pressure profile from sounding or modelling [hPa].}

\item{t_profile}{Temperature profile from sounding or modelling [K].}

\item{mr_profile}{Mixing ratio profile from sounding or modelling [kg/kg].}

\item{vtc}{logical refers is virtual temperature correction due to Doswell and
Rasmussen (1994).}
}
\value{
A vector containing CAPE, CIN, p_LCL, p_LFC
}
\description{
The function calculates the Convective Available Potential
 Energy (CAPE) using R. It returns a vector of CAPE, convective inhibition
 (CIN), pressure at lifted condensation level (LCL),and level of free
 convection (LFC).
}
\examples{
data("radiosonde")
t_profile <- radiosonde$temp + 273.15 # K
dwpt_profile <- radiosonde$dpt + 273.15 # K
p_profile <- radiosonde$pressure # hPa
# Mixing ratio
mr_profile <- mixing_ratio_from_dewpoint(dwpt_profile, p_profile)
cape_f(t_profile[1], dwpt_profile[1], mr_profile[1],
  p_profile, t_profile, mr_profile,
  vtc = TRUE
)
}
\author{
Klemens Barfus (Original in Fortran), Ahmed Homoudi (Integration in R)
}
