% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv2D_r.R
\name{conv2D_r}
\alias{conv2D_r}
\title{2D Convolution using R.}
\usage{
conv2D_r(a, b)
}
\arguments{
\item{a}{A matrix (2D array) of values.}

\item{b}{A matrix (2D array) of values.}
}
\value{
A matrix representing the 2D cross-correlation of the input matrices.
}
\description{
This function calculates the 2D Convolution of two matrices
 `a` and `b` using compiled Fortran subroutine.
}
\examples{
a <- matrix(c(1, 2, 3, 4), ncol = 2)
b <- matrix(c(5, 6, 7, 8), ncol = 2)
conv2D_f(a, b)
}
\author{
Ahmed Homoudi
}
