% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mixing_ratio_from_dewpoint}
\alias{mixing_ratio_from_dewpoint}
\title{Mixing ratio}
\usage{
mixing_ratio_from_dewpoint(dewpoint, pressure)
}
\arguments{
\item{dewpoint}{temperature [K]}

\item{pressure}{in [hPa]}
}
\description{
Estimation of mixing ratio from pressure [hPa] and dewpoint temperature [K].
}
\author{
Ahmed Homoudi
}
