% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiosonde.R
\docType{data}
\name{radiosonde}
\alias{radiosonde}
\title{Radiosonde data}
\format{
A data frame with 7 columns and 66 rows
}
\usage{
radiosonde
}
\description{
radiosonde example data to calculate CAPE.
The data measured at Pointe-Noire,  Republic of the Congo, radiosonde station
on 2024.02.02 12Z. The radiosonde data are available online by the University
of Wyoming \url{http://weather.uwyo.edu/upperair/sounding.html/}.
}
\keyword{datasets}
