% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{saturation_vapour_pressure}
\alias{saturation_vapour_pressure}
\title{Saturation vapour pressure}
\usage{
saturation_vapour_pressure(temperature, ice = FALSE)
}
\arguments{
\item{temperature}{in [k]}

\item{ice}{TRUE or FALSE, if to consider ice state or not.}
}
\description{
Estimation of Saturation vapour pressure [hPa] from temperature [k].
}
\author{
Klemens Barfus (Original in Fortran), translated to R by Ahmed Homoudi
}
