% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaFortR-package.R
\docType{package}
\name{AquaFortR-package}
\alias{AquaFortR}
\alias{AquaFortR-package}
\title{AquaFortR: Streamlining Atmospheric Science, Oceanography, Climate, and Water Research with Fortran-accelerated R}
\description{
AquaFortR is an educational package for students and researchers in Atmospheric Science, Oceanography, Climate, and Water Research. AquaFortR aims to show that using simple Fortran scripts can fulfil the need to speed up R, especially considering that most data sets in these fields are large discretised arrays representing earth system processes. Fortran is one of the fastest programming languages in number crunching, if not the fastest, and multidimensional arrays are a core part of Fortran.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://AHomoudi.github.io/AquaFortR}
  \item Report bugs at \url{https://github.com/AHomoudi/AquaFortR/issues}
}

}
\author{
\strong{Maintainer}: Ahmed Homoudi \email{ahmed.homoudi@tu-dresden.de} (\href{https://orcid.org/0000-0001-6906-5986}{ORCID})

}
\keyword{internal}
