% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiosonde.R
\docType{data}
\name{radiosonde}
\alias{radiosonde}
\title{Radiosonde data}
\format{
A data frame with 14 columns and 2962 rows
}
\source{
German Weather Service (DWD). (2024). Radiosondes data set.
Climate Data Center (CDC). \url{https://opendata.dwd.de/}
}
\usage{
radiosonde
}
\description{
radiosonde example data to calculate CAPE.
The data measured at Stuttgart radiosonde station on 2021-06-01 12:00 mid day.
For more information about about the
variables meaning, please refer to this sheet
\url{https://opendata.dwd.de/climate_environment/CDC/help/Abkuerzung_Spaltenname_CDC_20180914.xlsx}
}
\keyword{datasets}
